Modified to take CL arguments.However, there are some lingering import errors,
