+++ CONTRIBUTORS.rst
@@ -10,4 +10,4 @@ Development Lead
 Contributors
 ------------
 
-None yet. Why not be the first?
+* Neelay Shah <shahnh19gmail.com>
