+++ KD_Lib/original/model.py
@@ -1,0 +1,45 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class teacher(nn.Module):
+    def __init__(self,hidden_size):
+        super(teacher, self).__init__()
+
+        self.fc1 = nn.Linear(28*28, hidden_size)
+        self.fc2 = nn.Linear(hidden_size, hidden_size)
+        self.fc3 = nn.Linear(hidden_size, 10)
+        self.dropout = nn.Dropout2d(0.5)
+
+    def forward(self, x):
+        x = torch.flatten(x, 1)
+        x = self.fc1(x)
+        x = F.relu(x)
+        x = self.dropout(x)
+        x = self.fc2(x)
+        x = F.relu(x)
+        x = self.dropout(x)
+        out = self.fc3(x)
+
+        return out
+
+
+class student(nn.Module):
+
+    def __init__(self,hidden_size):
+        super(student, self).__init__()
+
+        self.fc1 = nn.Linear(28*28, hidden_size)
+        self.fc2 = nn.Linear(hidden_size, hidden_size)
+        self.fc3 = nn.Linear(hidden_size, 10)
+
+    def forward(self, x):
+        x = torch.flatten(x, 1)
+        x = self.fc1(x)
+        x = F.relu(x)
+        x = self.fc2(x)
+        x = F.relu(x)
+        out = self.fc3(x)
+
+        return out
