+++ tests/test_KD_Lib.py
@@ -5,6 +5,8 @@ import pytest
 import random
 
 from KD_Lib import KD_Lib
+from KD_Lib.KD_Lib.original.mnist import mnist
+
 
 
 @pytest.fixture
@@ -29,6 +31,9 @@ def test_max_number(generate_numbers):
 
     our_result = KD_Lib.max_number(generate_numbers)
     assert our_result == max(generate_numbers)
+
+def test_mnist():
+    mnist()
 
 
 # def test_max_number_bad(generate_numbers):
